% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneapi.R
\name{add_previous}
\alias{add_previous}
\title{add_previous}
\usage{
add_previous(d, nlead = 12)
}
\arguments{
\item{d}{with the variable of \code{Q_obs}}

\item{nlead}{the number of leads to add}
}
\description{
add_previous
}
