% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfold-package.R
\docType{package}
\name{kfold-package}
\alias{kfold}
\alias{kfold-package}
\title{kfold: Machine Learning for Runoff Prediction}
\description{
Machine learning In k-fold cross validation .
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com} (\href{https://orcid.org/0000-0003-1836-8172}{ORCID})

}
\keyword{internal}
