% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kford_ml.R
\name{kford_ml}
\alias{kford_ml}
\alias{kford_rf}
\alias{kford_xgboost}
\alias{kford_lm}
\title{kfold machine learning}
\usage{
kford_ml(X, Y, kfold = 5, FUN, ...)

kford_rf(X, Y, kfold = 5, FUN = ranger, ntree = 500, ...)

kford_xgboost(X, Y, kfold = 5, verbose = FALSE, nrounds = 500, ...)

kford_lm(X, Y, kfold = 5, ...)
}
\arguments{
\item{...}{optional parameters to be passed to the low level function
    \code{randomForest.default}.}

\item{ntree}{Number of trees to grow.  This should not be set to too
    small a number, to ensure that every input row gets predicted at
    least a few times. }

\item{verbose}{If 0, xgboost will stay silent. If 1, it will print information about performance.
If 2, some additional information will be printed out.
Note that setting \code{verbose > 0} automatically engages the
\code{cb.print.evaluation(period=1)} callback function.}

\item{nrounds}{max number of boosting iterations.}
}
\description{
kfold machine learning
}
\examples{
set.seed(1)
n <- 100 ; p <- 2
X <- matrix(rnorm(n * p), n, p) # no intercept!
y <- rnorm(n)

# r_rf  <- kford_rf(X, y)
r_lm  <- kford_lm(X, y)
r_xgb <- kford_xgboost(X, y)
}
