% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kford_rf.R
\name{kford_rf}
\alias{kford_rf}
\title{kfold randomForest}
\usage{
kford_rf(X, Y, kfold = 5, ntree = 500, ...)
}
\arguments{
\item{ntree}{Number of trees to grow.  This should not be set to too
    small a number, to ensure that every input row gets predicted at
    least a few times. }

\item{...}{optional parameters to be passed to the low level function
    \code{randomForest.default}.}
}
\description{
kfold randomForest
}
