% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kford_xgboost.R
\name{kford_xgboost}
\alias{kford_xgboost}
\title{kford xgboost}
\usage{
kford_xgboost(X, Y, kfold = 5, verbose = FALSE, nrounds = 500, ...)
}
\arguments{
\item{verbose}{If 0, xgboost will stay silent. If 1, it will print information about performance.
If 2, some additional information will be printed out.
Note that setting \code{verbose > 0} automatically engages the
\code{cb.print.evaluation(period=1)} callback function.}

\item{nrounds}{max number of boosting iterations.}

\item{...}{other parameters to pass to \code{params}.}
}
\description{
kford xgboost
}
