% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLrunoff-package.R
\docType{package}
\name{MLrunoff-package}
\alias{MLrunoff}
\alias{MLrunoff-package}
\title{MLrunoff: Machine Learning for Runoff Prediction}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com} (\href{https://orcid.org/0000-0003-1836-8172}{ORCID})

}
\keyword{internal}
