% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.kfold}
\alias{predict.kfold}
\title{predict for kfold object}
\usage{
\method{predict}{kfold}(object, newdata = NULL, ..., mode = "test")
}
\arguments{
\item{mode}{\itemize{
\item \code{train},
\item \code{test},
}}
}
\description{
predict for kfold object
}
