% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfold-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{plan}
\alias{multisession}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{future}{\code{\link[future]{multisession}}, \code{\link[future]{plan}}}
}}

