% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost.R
\name{train_xgboost}
\alias{train_xgboost}
\title{训练 XGB 洪水预报后处理模型}
\usage{
train_xgboost(data_full, leads = 1:12, ...)
}
\arguments{
\item{data_full}{训练数据, 含 \verb{site, time, Q_obs, P, Q_sim, PET_Romanenko}}

\item{leads}{预见期 (小时)}

\item{...}{透传给 \code{kfold_xgboost()}}
}
\value{
list: \code{data_full}, \code{data} (有效观测行), 及各模型族的 kfold 拟合结果
}
\description{
对 5 个模型族 (MetXGB / QlagXGB / HydroMetXGB / HydroMetQlagXGB) 分别做
kfold 训练. 特征由 \code{xgb_features()} 构造. 配套预测见 \code{\link[=predict_xgboost]{predict_xgboost()}},
表现检验见 \code{\link[=summary_xgboost]{summary_xgboost()}}.
}
