% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfold_calib.R
\name{kfold_calib}
\alias{kfold_calib}
\title{kfold_calib}
\usage{
kfold_calib(X, Y, FUN = xgboost, index = NULL, ..., ratio_valid = 0.3)
}
\arguments{
\item{index}{index of validation set}
}
\description{
kfold_calib
}
