% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kford_ml.R
\name{kford_ml}
\alias{kford_ml}
\alias{kford_rf}
\alias{kford_xgboost}
\alias{kford_lm}
\title{kfold machine learning}
\usage{
kford_ml(X, Y, kfold = 5, FUN, ...)

kford_rf(X, Y, kfold = 5, FUN = ranger, ntree = 500, importance = "none", ...)

kford_xgboost(X, Y, kfold = 5, verbose = FALSE, nrounds = 500, ...)

kford_lm(X, Y, kfold = 5, ...)
}
\arguments{
\item{...}{Further arguments passed to or from other methods (currently ignored).}

\item{importance}{Variable importance mode, one of 'none', 'impurity', 'impurity_corrected', 'permutation'. The 'impurity' measure is the Gini index for classification, the variance of the responses for regression and the sum of test statistics (see \code{splitrule}) for survival.}

\item{verbose}{Show computation status and estimated runtime.}

\item{nrounds}{max number of boosting iterations.}
}
\description{
kfold machine learning
}
\examples{
set.seed(1)
n <- 100 ; p <- 2
X <- matrix(rnorm(n * p), n, p) # no intercept!
y <- rnorm(n)

r_lm  <- kford_lm(X, y)
r_xgb <- kford_xgboost(X, y)
# r_rf  <- kford_rf(X, y)
}
\seealso{
\code{\link[ranger:ranger]{ranger::ranger()}}, \code{\link[xgboost:xgb.train]{xgboost::xgboost()}}
}
