% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/previous_tn.R
\name{previous_tn}
\alias{previous_tn}
\alias{previous_tn.default}
\alias{previous_tn.data.frame}
\title{previous_tn}
\usage{
previous_tn(x, n = 7, prefix = "", ...)

\method{previous_tn}{default}(x, n = 7, prefix = "", ...)

\method{previous_tn}{data.frame}(x, n = 7, ...)
}
\description{
previous_tn
}
\examples{
set.seed(1)
x <- rnorm(10)
previous_tn(x, 7, "R1_")
# data.frame
d = data.frame(x)
previous_tn(d)
}
